Autenticações aula 1 e 2 quase prontas

+++ meu-app-laravel/routes/web.php
@@ -1,32 +1,49 @@
 <?php
 
+use App\Http\Controllers\PostController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\{
-    PostController,
-    UserController,
-    TeamController,
-    ViaCepController
-};
 
-Route::get("/user/team/{id}", [TeamController::class, 'show'])->name('team.show');
+require __DIR__.'/auth.php';
+
+
+Route::get('/', function () {
+    return view('home');
+});
+
+
+Route::middleware(['auth'])->group(function () {
+    
+    Route::get("/user/team/{id}", [TeamController::class, 'show'])->name('team.show')->middleware('auth');
 //Route::get("/team/{id}", [TeamController::class, 'show'])->name('team.show');
 //Route::get("/user/team/{id}", [TeamController::class, 'show'])->name('team.team'); colocando o arquivo chamado team.blade dentro da pasta users
 
+    Route::get('/posts', [PostController::class, 'index'])->name('posts.index')->middleware('auth');
+    Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show')->middleware('auth');
 
-Route::get('/posts', [PostController::class, 'index'])->name('posts.index');
-Route::get('/users/{id}/posts', [PostController::class, 'show'])->name('posts.show');
+    Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('auth');
+    Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update')->middleware('auth');
+    Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware('auth');
+    Route::get('/users/create', [UserController::class, 'create'])->name('users.create')->middleware('auth');
+    Route::post('/user', [UserController::class, 'store'])->name('users.store')->middleware('auth');
+    Route::get('/users', [UserController::class, 'index'])->name('users.index')->middleware('auth');
+    Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show')->middleware('auth');
+});
 
-
-Route::delete('/users/{id}', [UserController::class, 'destroy'])->name('users.destroy');
-Route::put('/users/{id}', [UserController::class, 'update'])->name('users.update');
-Route::get('/users/{id}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::get('/users/create', [UserController::class, 'create'])->name('users.create');
-Route::post('/user', [UserController::class, 'store'])->name('users.store');
-Route::get('/users', [UserController::class, 'index'])->name('users.index');
-Route::get('/users/{id}', [UserController::class, 'show'])->name('users.show');
-
+Route::middleware(['auth','admin'])->group(function() {
+    Route::get('/admin',[UserController::class, 'admin'])->name('admin');
+});
 //VIA CEP WEB SERVICE
-Route::get('/viacep',[ViaCepController::class, 'index'])->name('viacep.index');
-Route::post('/viacep',[ViaCepController::class, 'index'])->name('viacep.index.post');
-Route::get('/viacep/{cep}', [ViaCepController::class, 'show'])->name('viacep.show');
+//Route::get('/viacep',[ViaCepController::class, 'index'])->name('viacep.index');
+//Route::post('/viacep',[ViaCepController::class, 'index'])->name('viacep.index.post');
+//Route::get('/viacep/{cep}', [ViaCepController::class, 'show'])->name('viacep.show');
+
+//CODIGO PARA CLICAR EM ESQUECI MINHA SENHA USAR NO .ENV
+//MAIL_MAILER=smtp
+//MAIL_HOST=smtp.gmail.com
+//MAIL_PORT=587
+//MAIL_USERNAME=risysintegrator@gmail.com
+//MAIL_PASSWORD=92392105
+//MAIL_ENCRYPTION=tls
+//MAIL_FROM_ADDRESS=risysintegrator@gmail.com
